Fixes quirk with first and last files in a directory
Vlc loops the playlist, so when there is no previous file (first
file in a directory) using Previous button would load the second
file from the directory. This changes the behavior by loading the
last file from the direcory as previous when there are more than
two files in a directory. The same goes for last file in a
directory and using Next button.